Combined Node and Edge editing into Map Editing. Done in code w/ me w/ Billy and Chris (for some strange reason Billys changes were merged with mine)
