+++ src/main/resources/edu/wpi/aquamarine_axolotls/fxml/EdgePopUp.fxml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import com.jfoenix.controls.JFXButton?>
+<?import com.jfoenix.controls.JFXComboBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.ColumnConstraints?>
+<?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.RowConstraints?>
+
+<GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.wpi.aquamarine_axolotls.views.mapping.EdgePopUp">
+  <columnConstraints>
+    <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+    <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+  </columnConstraints>
+  <rowConstraints>
+    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+  </rowConstraints>
+   <children>
+      <JFXComboBox fx:id="startNodeDropdown" prefHeight="25.0" prefWidth="329.0" promptText="Start Node" GridPane.rowIndex="1" />
+      <JFXComboBox fx:id="endNodeDropdown" prefHeight="25.0" prefWidth="355.0" promptText="End Node" GridPane.columnIndex="1" GridPane.rowIndex="1" />
+      <Label fx:id="edgeIDLabel" text="Label" />
+      <JFXButton onAction="#submit" text="Submit" GridPane.rowIndex="2" />
+      <HBox prefHeight="100.0" prefWidth="200.0" GridPane.columnIndex="1" GridPane.rowIndex="2">
+         <children>
+            <JFXButton onAction="#delete" text="Delete" />
+            <Label fx:id="submissionStatusLabel" />
+         </children>
+      </HBox>
+   </children>
+</GridPane>
